Move file transfer tests into module
